requirements.txt
absl-py==1.0.0
altair==4.2.0
annotated-types==0.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
blinker==1.4
blis==0.7.11
cachetools==5.0.0
catalogue==2.0.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.8
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
entrypoints==0.4
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl
executing==0.8.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.0
flatbuffers==2.0
fonttools==4.29.1
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl
fsspec==2024.9.0
gast==0.5.3
gekko==1.0.2
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
huggingface-hub==0.24.7
idna==3.9
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
Janome==0.5.0
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.4.0
language_data==1.2.0
Levenshtein==0.25.1
libclang==13.0.0
lxml==4.7.1
marisa-trie==1.2.0
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mpmath==1.3.0
mrjob==0.7.4
murmurhash==1.0.10
mypy-extensions==0.4.3
nano==0.10.0
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
notebook==6.4.8
numpy==1.26.4
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
plotly==5.5.0
preshed==3.0.9
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.7.1
Pygments==2.18.0
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-Levenshtein==0.25.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.1
PyYAML==6.0
pyzmq==22.3.0
rapidfuzz==3.9.7
regex==2024.5.15
requests==2.27.1
requests-file==2.1.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.8
safetensors==0.4.5
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.0
soupsieve==2.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
srsly==2.4.8
stack-data==0.1.4
statsmodels==0.13.1
streamlit==1.5.0
sympy==1.13.2
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
Theano==1.0.5
thinc==8.2.5
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
torch==2.4.1
tornado==6.1
tqdm==4.66.4
traitlets==5.1.1
transformers==4.44.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
vboxapi==1.0
wasabi==1.1.3
watchdog==2.1.6
wcwidth==0.2.5
weasel==0.4.1
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
xgboost==1.5.2
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.7.0/zh_core_web_sm-3.7.0-py3-none-any.whl
